+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `kt-rules:no-since-in-kdoc` won't indent by an extra space while adding `@since` to a top-level KDoc
 - `kt-rules:no-since-in-kdoc` won't add `@since null` to a KDoc if the current version property is `null`
+- `kt-rules:kdoc-content-wrapping` won't delete KDoc ends (`*/`) from inside code blocks while wrapping
 
 ## [1.1.1] - 2023-04-29